requirements.txt
Django==1.7.1
Pillow==2.5.1
PyYAML==3.11
argh==0.25.0
django-bower==5.0.1
django-admin-sortable2==0.3.1
django-ckeditor==4.4.6
django-endless-pagination==2.0
django-extensions==1.3.8
django-lineage==0.2.0
django-sekizai==0.7
django-shell-plus==1.1.5
gunicorn==19.0.0
meld3==1.0.0
pathtools==0.1.2
psycopg2==2.5.3
pytz==2014.4
six==1.7.3
watchdog==0.8.0
wsgiref==0.1.2
django-supervisor==0.4.0